pyproject.toml
[project]
name = "my-adventures-with-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "datasets>=3.2.0",
    "gensim>=4.3.3",
    "huggingface>=0.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "nltk>=3.9.1",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.18.0",
    "tokenizers>=0.21.0",
    "torchmetrics>=1.6.1",
    "torchvision>=0.20.1",
    "transformers>=4.48.1",
    "wandb>=0.19.4",
]
